feat: auth setup, dev mode and general refactoring
