+++ justfile
@@ -74,65 +74,21 @@ check-all: check clippy tsc lint
 fmt:
     cd app && bunx prettier --write src/
 
-# === TUI Development ===
+# === TUI ===
 
-# Initialize TUI dev database (creates isolated database)
-init-tui-db:
-    @echo "🔧 Creating isolated dev database for TUI..."
-    @echo "⚠️  This will NOT touch your production database!"
-    PGPASSWORD=password createdb -U postgres -h localhost toki_tui_dev || echo "Database already exists (that's OK)"
-    @echo "📦 Running migrations on toki_tui_dev..."
-    cd toki-api && DATABASE_URL=postgres://postgres:password@localhost:5432/toki_tui_dev sqlx migrate run
-    @echo "👤 Creating test user..."
-    PGPASSWORD=password psql -U postgres -h localhost -d toki_tui_dev -c "INSERT INTO users (email, full_name, picture, access_token, roles) VALUES ('test@example.com', 'Test User', 'https://example.com/avatar.jpg', 'test_token', ARRAY['User']::text[]) ON CONFLICT (email) DO NOTHING;" > /dev/null
-    @echo "✅ TUI dev database ready!"
-
-# Run the TUI (with dev database)
+# Run the TUI (requires login — run `just tui-login` first if needed)
 tui:
     cd toki-tui && cargo run
 
-# Reset TUI dev database (clean slate for testing)
-reset-tui-db:
-    @echo "🗑️  Dropping TUI dev database..."
-    PGPASSWORD=password dropdb -U postgres -h localhost toki_tui_dev --if-exists
-    @echo "🔧 Recreating..."
-    just init-tui-db
+# Run the TUI in dev mode (no login required, mock data)
+tui-dev:
+    cd toki-tui && cargo run -- --dev
 
-# Check which databases exist (verify isolation)
-check-dbs:
-    @echo "📊 Existing databases:"
-    PGPASSWORD=password psql -U postgres -h localhost -c "SELECT datname FROM pg_database WHERE datname LIKE 'toki%';"
+# Authenticate the TUI via browser OAuth
+tui-login:
+    cd toki-tui && cargo run -- --login
 
-# Check timer entries in TUI dev database
-check-timers:
-    @echo "📋 Timer entries in dev database:"
-    @echo ""
-    @echo "Active timers:"
-    @PGPASSWORD=password psql -U postgres -h localhost -d toki_tui_dev -c "SELECT id, start_time, project_name, activity_name, note FROM timer_history WHERE end_time IS NULL;" || echo "No active timers"
-    @echo ""
-    @echo "Recent timer history (last 10):"
-    @PGPASSWORD=password psql -U postgres -h localhost -d toki_tui_dev -c "SELECT id, start_time, end_time, project_name, activity_name, note FROM timer_history ORDER BY start_time DESC LIMIT 10;"
+# Log out (clear saved session)
+tui-logout:
+    cd toki-tui && cargo run -- --logout
 
-# Open interactive database shell for TUI dev database
-db-shell:
-    @echo "Opening database shell for toki_tui_dev..."
-    @echo "Useful commands:"
-    @echo "  \dt              - List tables"
-    @echo "  \d timer_history - Describe timer_history table"
-    @echo "  \q               - Quit"
-    @echo ""
-    PGPASSWORD=password psql -U postgres -h localhost -d toki_tui_dev
-
-# Create test timer history entries for demo
-create-test-timers:
-    @echo "📝 Creating test timer history entries..."
-    @PGPASSWORD=password psql -U postgres -h localhost -d toki_tui_dev -c \
-        "INSERT INTO timer_history (user_id, start_time, end_time, project_name, activity_name, note) VALUES \
-        (1, NOW() - INTERVAL '5 hours', NOW() - INTERVAL '3 hours', 'Toki2 Development', 'Backend Development', 'Fixed timer persistence bug'), \
-        (1, NOW() - INTERVAL '8 hours', NOW() - INTERVAL '6 hours', 'TUI Development', 'Feature Implementation', 'Added project selection UI'), \
-        (1, NOW() - INTERVAL '2 days', NOW() - INTERVAL '2 days' + INTERVAL '90 minutes', 'Azure DevOps Integration', 'API Integration', 'Implemented PR webhooks'), \
-        (1, NOW() - INTERVAL '3 days', NOW() - INTERVAL '3 days' + INTERVAL '45 minutes', 'Internal Tools', 'Documentation', 'Updated setup guide'), \
-        (1, NOW() - INTERVAL '1 day', NOW() - INTERVAL '1 day' + INTERVAL '2 hours', 'TUI Development', 'UI Design', 'Designed history view layout') \
-        ON CONFLICT DO NOTHING;"
-    @echo "✅ Created 5 test timer entries"
-    @just check-timers
