--- toki-tui/docs/TUI_FEATURE_SUMMARY.md
@@ -1,297 +1,0 @@
-# TUI Feature Update Summary
-
-## ✅ What Was Implemented
-
-### 1. **Project & Activity Selection**
-- **5 predefined test projects** with unique activities for each
-- **Interactive selection UI** with arrow key navigation
-- **Keyboard shortcuts**: `P` for projects, `A` for activities
-- **Visual feedback**: Selected item highlighted in yellow
-- Projects include codes (e.g., [TOKI-2], [TUI])
-
-### 2. **Timer History View**
-- **See all past timers** in chronological order (newest first)
-- **Shows**: Date, time, duration, project, activity, note
-- **Keyboard shortcuts**: `H` or `Tab` to view, `Tab` to return
-- **Scrollable**: Up/down arrows to browse
-- **Database query**: Fetches last 50 entries
-
-### 3. **Enhanced Timer View**
-- **Shows selected project/activity** even when timer is stopped
-- **Uses selected values** when starting new timer
-- **Status messages** for all actions
-- **Improved controls display** with all available shortcuts
-
----
-
-## 🎯 How It Works
-
-### Architecture
-
-```
-toki-tui/
-├── src/
-│   ├── test_data.rs         # ✨ NEW: Test projects & activities
-│   ├── app.rs               # ✨ UPDATED: Multi-view support + selection logic
-│   ├── ui/mod.rs            # ✨ UPDATED: 4 separate view renderers
-│   ├── api/database.rs      # ✨ UPDATED: Added get_timer_history()
-│   └── main.rs              # ✨ UPDATED: New keyboard shortcuts
-```
-
-### State Management
-
-**Views**: 
-- `Timer` - Main timer display (default)
-- `History` - Past timer entries
-- `SelectProject` - Project picker
-- `SelectActivity` - Activity picker (filtered by selected project)
-
-**Selection State**:
-- `selected_project` - Currently chosen project
-- `selected_activity` - Currently chosen activity
-- `selected_project_index` - For UI highlighting
-- `selected_activity_index` - For UI highlighting
-
-### Navigation Flow
-
-```
-Timer View
-  ├─ Press P ──> Project Selection ──> Select ──> Back to Timer
-  ├─ Press A ──> Activity Selection ──> Select ──> Back to Timer
-  └─ Press H/Tab ──> History View ──> Press Tab ──> Back to Timer
-```
-
----
-
-## 🎮 Keyboard Shortcuts
-
-### Timer View
-| Key | Action |
-|-----|--------|
-| `Space` | Start/stop timer |
-| `P` | Select project |
-| `A` | Select activity |
-| `H` or `Tab` | View history |
-| `R` | Refresh from database |
-| `Q` | Quit |
-
-### Project/Activity Selection
-| Key | Action |
-|-----|--------|
-| `↑↓` | Navigate list |
-| `Enter` | Confirm selection |
-| `Esc` | Cancel (return to timer) |
-| `Q` | Quit |
-
-### History View
-| Key | Action |
-|-----|--------|
-| `↑↓` | Scroll entries |
-| `Tab` | Return to timer |
-| `Q` | Quit |
-
----
-
-## 📦 Test Data
-
-### Projects (5 total)
-1. **Toki2 Development** [TOKI-2]
-   - Backend Development, Frontend Development, Bug Fixes, Code Review
-2. **Azure DevOps Integration** [ADO-INT]
-   - API Integration, Webhook Setup, Testing
-3. **TUI Development** [TUI]
-   - UI Design, Feature Implementation, Testing & Debugging
-4. **Internal Tools** [TOOLS]
-   - Development, Maintenance, Documentation
-5. **Research & Learning**
-   - Learning, Experimentation, Proof of Concept
-
-### How to Create Test History Entries
-```bash
-just create-test-timers
-```
-
-This creates 5 sample completed timers spanning several days.
-
----
-
-## 🚀 Usage Example
-
-1. **Launch TUI**:
-   ```bash
-   just tui
-   ```
-
-2. **Select Project**:
-   - Press `P`
-   - Arrow down to "TUI Development"
-   - Press `Enter`
-
-3. **Select Activity**:
-   - Press `A`
-   - Arrow down to "Feature Implementation"
-   - Press `Enter`
-
-4. **Start Timer**:
-   - Press `Space`
-   - Timer starts counting!
-
-5. **View History**:
-   - Press `H`
-   - See your completed timers
-   - Press `Tab` to return
-
----
-
-## 🔧 Technical Details
-
-### New Database Function
-```rust
-pub async fn get_timer_history(&self, user_id: i32, limit: i64) -> Result<Vec<TimerHistoryEntry>>
-```
-Fetches timer history for display, ordered by most recent first.
-
-### New Data Structures
-```rust
-pub struct TestProject {
-    pub id: String,
-    pub name: String,
-    pub code: Option<String>,
-}
-
-pub struct TestActivity {
-    pub id: String,
-    pub name: String,
-    pub project_id: String,
-}
-```
-
-### View-Specific Rendering
-Each view has its own render function:
-- `render_timer_view()` - Main timer interface
-- `render_history_view()` - Scrollable history list
-- `render_project_selection()` - Project picker
-- `render_activity_selection()` - Activity picker
-
----
-
-## 🎨 UI Improvements
-
-### Timer View
-- Shows selected project/activity even when stopped
-- Format: "Project / Activity (not running)"
-- When running: "Project / Activity"
-
-### Selection Views
-- **Highlighted selection**: Yellow bold text
-- **Project codes**: Displayed as [CODE] before name
-- **Auto-return**: Selecting an item returns to timer view
-- **Esc support**: Cancel anytime
-
-### History View
-- **Date/time formatting**: "2026-02-11 07:22"
-- **Duration display**: "02:30" (HH:MM)
-- **Active indicators**: Shows "Active" for running timers
-- **Entry count**: Shows total in title
-
----
-
-## ⚠️ Current Limitations
-
-1. **Stop = Delete**: Stopping a timer deletes it (doesn't save to Milltime yet)
-2. **No note editing**: Timer note is hardcoded to "Working on task"
-3. **No history scrolling UI**: Only up/down keys (no scroll bar)
-4. **Activities reset**: Selecting a new project resets activity to first item
-5. **No statistics**: No totals or aggregations
-
----
-
-## 🎯 Testing Commands
-
-```bash
-# Run the TUI
-just tui
-
-# Check timer database
-just check-timers
-
-# Create test history data
-just create-test-timers
-
-# Reset database (clean slate)
-just reset-tui-db
-
-# Open database shell
-just db-shell
-```
-
----
-
-## 📊 Files Modified
-
-### New Files
-- `toki-tui/src/test_data.rs` - Test projects and activities
-- `TEST_TUI_FEATURES.md` - Detailed testing guide
-
-### Modified Files
-- `toki-tui/src/app.rs` - Added multi-view support, selection logic
-- `toki-tui/src/ui/mod.rs` - Completely rewritten for 4 views
-- `toki-tui/src/api/database.rs` - Added get_timer_history()
-- `toki-tui/src/main.rs` - Enhanced event handling, new shortcuts
-- `justfile` - Added create-test-timers command
-
-### Lines Changed
-- ~500 lines of new code
-- ~200 lines modified
-- 4 new views implemented
-- 8 new keyboard shortcuts
-
----
-
-## 🎉 Success Criteria
-
-✅ Can select from 5 different projects  
-✅ Can select activities specific to each project  
-✅ Timer uses selected project/activity when started  
-✅ Selected project/activity persists when timer stops  
-✅ History view shows past timers with all details  
-✅ Navigation between views works smoothly  
-✅ All keyboard shortcuts functional  
-✅ UI renders correctly in terminal  
-✅ Database persistence works  
-✅ Test data can be created easily  
-
-**All features complete and tested!** 🚀
-
----
-
-## 🔮 Next Steps (Future)
-
-### Priority 1 (Next Session)
-- [ ] Note editing with input dialog
-- [ ] Save timer to Milltime instead of deleting
-- [ ] Visual indicator for active timer in history
-
-### Priority 2 (Soon)
-- [ ] Pagination for long history
-- [ ] Statistics view (totals per project)
-- [ ] Edit timer after starting
-
-### Priority 3 (Later)
-- [ ] Real Milltime API integration
-- [ ] Authentication (session cookies)
-- [ ] Git integration for notes
-- [ ] Export to CSV
-
----
-
-## 📚 Documentation
-
-- **Testing Guide**: `TEST_TUI_FEATURES.md`
-- **Setup Guide**: `toki-tui/SETUP_GUIDE.md`
-- **Database Cheatsheet**: `DATABASE_CHEATSHEET.md`
-- **Project README**: `toki-tui/README.md`
-
----
-
-**Ready to use!** Run `just tui` to try it out. 🎮
