+++ toki-tui/src/api/mod.rs
@@ -1,3 +1,2 @@
-pub mod database;
-
-pub use database::Database;
+// API module — HTTP client replaces direct DB access.
+// TimerHistoryEntry is now in crate::types.
