+++ toki-tui/src/ui/utils.rs
@@ -1,0 +1,31 @@
+use ratatui::layout::{Constraint, Direction, Layout, Rect};
+use time::UtcOffset;
+
+pub fn to_local_time(dt: time::OffsetDateTime) -> time::OffsetDateTime {
+    if let Ok(local_offset) = UtcOffset::current_local_offset() {
+        dt.to_offset(local_offset)
+    } else {
+        dt
+    }
+}
+
+/// Helper function to create a centered rectangle
+pub fn centered_rect(width: u16, height: u16, r: Rect) -> Rect {
+    let popup_layout = Layout::default()
+        .direction(Direction::Vertical)
+        .constraints([
+            Constraint::Length((r.height.saturating_sub(height)) / 2),
+            Constraint::Length(height),
+            Constraint::Length((r.height.saturating_sub(height)) / 2),
+        ])
+        .split(r);
+
+    Layout::default()
+        .direction(Direction::Horizontal)
+        .constraints([
+            Constraint::Length((r.width.saturating_sub(width)) / 2),
+            Constraint::Length(width),
+            Constraint::Length((r.width.saturating_sub(width)) / 2),
+        ])
+        .split(popup_layout[1])[1]
+}
